+++ backend/inventarwerk_api/static/index.html
